keyless/immutable test generalization
